requirements.txt
flask>=1.0.2
tensorflow>=1.10.1
numpy>=1.15.4
pillow>=4.0.0
keras>=2.2.2

